requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv>=1.0.0
supabase>=2.3.0
openai>=1.3.0
google-generativeai>=0.8.0
anthropic>=0.7.0
pydantic>=2.0.0
python-multipart>=0.0.6
numpy>=1.24.0
PyPDF2>=3.0.0
python-docx>=0.8.11
